Renamed data_attributes.py to data_profile.py
